create abstract class for TaskManager tests
